fix(codeql): safe_log + int-cast am Motor-Drive-Logger
CodeQL fand 2 Alerts an robot/server.py:392 (beide log-injection
medium): einen pro getainteten Argument-Pfad (request.direction
und request.speed * 100). Der vorherige Versuch via Pydantic
Literal hatte zwar request.direction strukturell eingegrenzt, aber
CodeQL trackt Literal-Constraints nicht als Sanitizer.

Workaround:
- safe_log(request.direction) bricht den direction-Pfad (delete-
  Pattern, von CodeQL als log-injection Sanitizer anerkannt).
- int(request.speed * 100) bricht den speed-Pfad. CodeQL erkennt
  primitive Type-Casts (int/str/float) bei Float-Eingaben als
  Narrowing -> kein Taint-Flow mehr in den logger-Call.
- Format-String von %.0f%% auf %d%% gewechselt (passt zum
  int-Argument; visuelle Ausgabe identisch).

Die Pydantic-Literal-Constraint aus dem vorherigen Commit bleibt
drin -- ist eine ECHTE Defense-in-Depth (422 fuer invalid Werte
am API-Boundary), unabhaengig davon dass CodeQL sie nicht als
Sanitizer erkennt.

Tests: 66 passed in robot_server_harmony + robot_protocol;
4986 erwartet im full run (unveraendert).

+++ src/elder_berry/robot/server.py
@@ -385,11 +385,14 @@ class RobotServer:
         @self.app.post("/motor/drive")
         def drive(request: DriveRequest) -> dict:
             self._motors.drive(request.direction, request.speed)
-            # direction ist Literal-constrained -> kein log-injection
-            # Vektor mehr; speed ist float und kann keine CR/LF tragen.
+            # safe_log + int-Cast bricht den Taint-Flow fuer beide
+            # Felder. Pydantic Literal allein reicht CodeQL nicht --
+            # request.direction wird trotz Literal-Constraint noch als
+            # user-provided getrackt, und request.speed * 100 (float)
+            # gilt CodeQL ebenfalls als getainted.
             logger.info(
-                "Motor: %s @ %.0f%%",
-                safe_log(request.direction), request.speed * 100,
+                "Motor: %s @ %d%%",
+                safe_log(request.direction), int(request.speed * 100),
             )
             resp = ApiResponse(
                 success=True,
